+++ CHANGELOG.md
@@ -9,6 +9,17 @@
 
 - Technical :
 
+## 5.6.0
+- Features:
+  - Hide session ID in streamer mode
+  - Night order in the side menu for both players and storyteller
+  - Join/Host buttons in side menu when disconnected
+  - Storyteller controlles for pregame in side menu (choose edition, add players, pick roles, send roles)
+  - Soundboard in storyteller sidemenu (ringing, rooster, gavel)
+  - Secret votes option toggleable from side menu for storyteller
+- Bugfixes:
+  - [P] Hotkey for adding multiple players
+  - Missing message for add multiple players prompt
 
 ## 5.5.0
 - Features: